fix: - pass licenser fields to services - fix organisation_member_test.go & organisation_test.go

+++ database/postgres/organisation_member_test.go
@@ -91,7 +91,7 @@ func TestCountOrganisationMembers(t *testing.T) {
 	memberCount, err := organisationMemberRepo.CountOrganisationMembers(context.Background())
 
 	require.NoError(t, err)
-	require.Equal(t, count, memberCount)
+	require.Equal(t, int64(count), memberCount)
 }
 
 func TestLoadUserOrganisationsPaged(t *testing.T) {
